LayoutTests/platform/gtk/http/tests/inspector/network/resource-sizes-memory-cache-expected.txt

Test for Resource size values (transfer size, decoded size, header size, request and response) for resources served from the memory cache.


== Running test suite: Resource.Size.MemoryCache
-- Running test case: Resource.Size.MemoryCache.Basic
PASS: statusCode should be 304.
PASS: compressed should be false.
PASS: responseSource should be Symbol(memory-cache).
PASS: estimatedNetworkEncodedSize should be exactly 0 bytes.
PASS: estimatedTotalTransferSize should be >= (encoded body size + headers).
PASS: size should be exactly 27 bytes.
FAIL: networkEncodedSize should be exactly 0 bytes.
    Expected: 0
    Actual: NaN
FAIL: networkTotalTransferSize should be >= (encoded body size + headers).
    Expected: greater than or equal to 20
    Actual: NaN
FAIL: requestBodyTransferSize should be exactly 0 bytes.
    Expected: 0
    Actual: NaN
FAIL: responseBodyTransferSize should be exactly 0 bytes.
    Expected: 0
    Actual: NaN
FAIL: requestHeadersTransferSize should be non-empty.
    Expected: greater than 20
    Actual: NaN
FAIL: responseHeadersTransferSize should be non-empty.
    Expected: greater than 20
    Actual: NaN
